+++ packages/mel-components/src/mel-select/src/types.ts
@@ -25,44 +25,50 @@ export interface MelSelectSlots {
   /**
    * 默认插槽, 可自定义渲染 option 组件列表
    */
-  default?: () => void;
+  'default'?: () => void;
   /**
    * 下拉列表顶部的内容
    * @version element-plus@2.4.3
    */
-  header?: () => void;
+  'header'?: () => void;
   /**
    * 下拉列表底部的内容
    * @version element-plus@2.4.3
    */
-  footer?: () => void;
+  'footer'?: () => void;
   /**
    * Select 组件头部内容
    */
-  prefix?: () => void;
+  'prefix'?: () => void;
   /**
    * 无选项时的列表
    */
-  empty?: () => void;
+  'empty'?: () => void;
   /**
    * select 组件自定义标签内容
    * @version element-plus@2.5.0
    */
-  tag?: () => void;
+  'tag'?: () => void;
   /**
    * select 组件自定义 loading 内容
    * @version element-plus@2.5.2
    */
-  loading?: () => void;
+  'loading'?: () => void;
   /**
    * select 组件自定义标签内容
    * @version element-plus@2.7.4
    */
-  label?: () => void;
+  'label'?: () => void;
   /**
    * 自定义渲染单个选项
    * @param option 选项数据
    * @version @mixte/mel-components@3.3.0
    */
-  option?: (option: MelSelectOption) => void;
+  'option'?: (option: MelSelectOption) => void;
+  /**
+   * 自定义渲染选项内容 ( 仅内容部分，而非整个选项 )
+   * @param option 选项数据
+   * @version @mixte/mel-components@3.3.0
+   */
+  'option-label'?: (option: MelSelectOption) => void;
 }